refactor(web): extract backtesting configuration data
